image-ai/requirements.txt
tensorflow==2.15.0.post1
matplotlib==3.8.3
pydot==2.0.0
pillow==10.2.0
numpy==1.26.4
sanic==23.12.1
